LayoutDiffusion/requirements.txt
# The following can be installed with pip using 'pip install -r requirements.txt'
# Make sure you have upgraded pip
# pip install --upgrade pip

pillow
einops
spacy==3.2.4
huggingface_hub==0.4.0
typing-inspect==0.8.0
typing_extensions==4.5.0
scipy==1.7.1
pytorch_fid==0.2.0
seaborn==0.11.2
pycocotools==2.0.2